Implmented naive linq solution
This first loops over the input to find a the max value in the
collection.
Then generates a range of numbers from 0 to our max value.
Finally we performn an linq Except to find the missing number.

I assume the .Except to be O(N²) worst case as it will loop over
the first collection and then compare this value to each value in
the second collection until a match is found or the collection is
exhausted.

A subtle nicety of IEnumerable and linq here is that this is
all lazy eveluated (after the inital .Max() call).
So Enumerable.Range will generate a `0` we then start
comparing this with each value in nums array,
if we find that it is missing the value is returned breaking
out of the loops.