Handle some edge cases
When the largest number is removed from the collection:
To solve this I updated the + 1 in the Enumerable.Rage to + 2.
Since this last value will only be reached if a missing number is not
found before then there is no adverse side effects extending this
range for all cases.

When the collection is empty:
Check if the nums.Length is > 0, if it is not .Max() will throw
so we can use 0 in this case.