Updated to a bitwise xor solution.
If we were to xor each number in the sequence with itself - we would
get 0 as the result.
We can leverage this to find the missing number, if we xor every value
in the collection along with a collection not missing any values.
the remainder would be the missing value.

Starting with the nums.Length as a seed (as this is the largest number
that should be in the collection).
We can then xor this with every value, and use `i` from the for loop in
place of a second list not missing any values.
Since we are reducing all the values into a single number the order
we do this is not important, so we dont need to worry that the list is not
sorted and that `i` will not nessarly match the value `nums[i]`.